fix: search results on url params defined
