+++ apps/watch/src/components/VideosPage/VideosPage.tsx
@@ -3,8 +3,9 @@ import Box from '@mui/material/Box'
 import Container from '@mui/material/Container'
 import Divider from '@mui/material/Divider'
 import Stack from '@mui/material/Stack'
-import type { ReactElement, ReactNode } from 'react'
-import { Index, useConfigure } from 'react-instantsearch'
+import { useRouter } from 'next/compat/router'
+import { type ReactElement, type ReactNode, useEffect } from 'react'
+import { Index, useConfigure, useInstantSearch } from 'react-instantsearch'
 
 import { GET_LANGUAGES } from '@core/journeys/ui/useLanguagesQuery'
 import { ThemeMode } from '@core/shared/ui/themes'
@@ -21,8 +22,32 @@ interface VideosPageProps {
   index?: boolean
 }
 
+interface FilterParams {
+  query: string | null
+  languageId: string | null
+  subtitleId: string | null
+}
+
+function extractQueryParams(url: string): FilterParams {
+  const params = new URLSearchParams(url.split('?')[1])
+  const query = params.get('query')
+  const languageId = params.get('menu[languageId]')
+  const subtitleId = params.get('menu[subtitles]')
+  return { query, languageId, subtitleId }
+}
+
 export function VideosPage({ index = false }: VideosPageProps): ReactElement {
   const indexName = process.env.NEXT_PUBLIC_ALGOLIA_INDEX ?? ''
+
+  const router = useRouter()
+  const decodedUrl = decodeURIComponent(router?.asPath ?? '')
+  const { query, languageId, subtitleId } = extractQueryParams(decodedUrl)
+  const hasQueryParams = !(query == null && languageId == null && subtitleId == null)
+
+  const { refresh } = useInstantSearch()
+  useEffect(() => {
+    if (hasQueryParams) refresh()
+  }, [])
 
   const { data: languagesData, loading: languagesLoading } =
     useQuery<GetLanguages>(GET_LANGUAGES, {
